v0.7.2: Final consolidation and performance refinements
🔧 Final Consolidation and Performance Refinements

Added:
- Comprehensive refactor plan for future database infrastructure extraction to shit/db/
- Enhanced logging configuration with reduced third-party library verbosity
- Improved dry-run consistency across all database operations

Changed:
- Database consolidation refinements with better error handling
- Logging verbosity reduction (default INFO → WARNING)
- SQLAlchemy engine configuration (disabled SQL echo)
- Third-party library logging (sqlalchemy, boto3, botocore, urllib3, aiosqlite)

Technical Improvements:
- Consistent dry-run behavior using unified processing method
- Better error isolation for individual S3 record processing
- Optimized database operations (reduced logging frequency)
- Enhanced CLI consistency (moved verbose/log-level to subcommands)

Performance Optimizations:
- Reduced log noise from third-party libraries
- Faster processing with optimized S3 existence checks
- Better resource management and connection pooling
- Memory efficiency improvements

Documentation:
- Created comprehensive refactor plan (reference_docs/shitvault_refactor_plan.md)
- Architecture documentation for current vs target state
- Step-by-step migration strategy for database infrastructure

Files Modified:
- shitvault/cli.py - Enhanced logging and argument parsing
- shitvault/shitpost_db.py - Performance optimizations and logging
- reference_docs/shitvault_refactor_plan.md - Comprehensive refactor planning
- CHANGELOG.md - Updated to v0.7.2

Benefits:
- Production-ready logging with clean, focused output
- Better user experience with consistent CLI behavior
- Future-ready architecture with clear refactoring path
- Enhanced maintainability and error handling
- Improved performance with optimized operations
