+++ CHANGELOG.md
@@ -17,6 +17,53 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Advanced confidence scoring and quality metrics
 - Batch processing optimizations
 - Real-time streaming analysis
+- Database infrastructure refactor (see `reference_docs/shitvault_refactor_plan.md`)
+
+---
+
+## [0.7.2] - 2025-01-15
+
+### 🔧 **Final Consolidation and Performance Refinements**
+
+#### Added
+- **Comprehensive refactor plan** - Created detailed plan for future database infrastructure extraction to `shit/db/`
+- **Enhanced logging configuration** - Reduced third-party library verbosity for cleaner output
+- **Improved dry-run consistency** - Unified dry-run output format across all database operations
+
+#### Changed
+- **Database consolidation refinements** - Further optimized S3 → Database processing with better error handling
+- **Logging verbosity reduction** - Changed default log level from INFO to WARNING for cleaner output
+- **SQLAlchemy engine configuration** - Disabled SQL echo for cleaner database operation logs
+- **Third-party library logging** - Reduced verbosity for sqlalchemy, boto3, botocore, urllib3, and aiosqlite
+
+#### Technical Improvements
+- **Consistent dry-run behavior** - Both dry-run and regular modes now use same processing method with internal dry-run handling
+- **Better error isolation** - Individual S3 record processing errors don't stop entire batch
+- **Optimized database operations** - Reduced logging frequency for successful operations (every 500 vs 100 records)
+- **Enhanced CLI consistency** - Moved verbose and log-level arguments to subcommands for proper parsing
+
+#### Performance Optimizations
+- **Reduced log noise** - Third-party libraries now log at WARNING/ERROR level only
+- **Faster processing** - Optimized S3 existence checks and database operations
+- **Better resource management** - Improved connection pooling and cleanup procedures
+- **Memory efficiency** - Streamlined data transformation and storage operations
+
+#### Documentation
+- **Refactor planning** - Created comprehensive plan for future `shit/db/` infrastructure extraction
+- **Architecture documentation** - Detailed analysis of current state vs target state for refactoring
+- **Migration strategy** - Step-by-step plan for extracting base database components
+
+#### Files Modified
+- `shitvault/cli.py` - Enhanced logging configuration and argument parsing
+- `shitvault/shitpost_db.py` - Performance optimizations and logging improvements
+- `reference_docs/shitvault_refactor_plan.md` - Comprehensive refactor planning document
+
+#### Benefits
+- **Production-ready logging** - Clean, focused output suitable for production environments
+- **Better user experience** - Consistent CLI behavior and clearer output
+- **Future-ready architecture** - Clear path for database infrastructure refactoring
+- **Enhanced maintainability** - Better separation of concerns and error handling
+- **Improved performance** - Optimized operations and reduced overhead
 
 ---
